Moved  layer_database file to keras package
